facts-search/requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dataclasses==0.6
faiss-cpu==1.6.4.post2
filelock==3.0.12
future==0.18.2
idna==2.10
joblib==0.17.0
nltk==3.5
numpy==1.19.4
packaging==20.4
pyparsing==2.4.7
regex==2020.11.13
requests==2.25.0
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
sentence-transformers==0.3.8
sentencepiece==0.1.94
six==1.15.0
threadpoolctl==2.1.0
tokenizers==0.8.1rc2
torch==1.7.0
tqdm==4.51.0
transformers==3.3.1
typing-extensions==3.7.4.3
urllib3==1.26.2
